Wrapper works for cmd objects, dangling link test works

+++ src/updater/main.c
@@ -262,7 +262,9 @@ static int find_deps(struct graph *g, struct node *n)
 	struct flist f;
 	char object_dir[] = ".tup/object/" SHA1_XD;
 	char cmdfile[] = ".tup/object/" SHA1_XD "/.cmd";
+	char depfile[] = ".tup/object/" SHA1_XD "/.secondary";
 	struct stat st;
+	struct stat st2;
 
 	tupid_to_xd(object_dir + 12, n->tupid);
 	flist_foreach(&f, object_dir) {
@@ -275,10 +277,14 @@ static int find_deps(struct graph *g, struct node *n)
 		}
 		if(n->node == NODE_FILE) {
 			tupid_to_xd(cmdfile+12, f.filename);
+			tupid_to_xd(depfile+12, f.filename);
 			if(stat(cmdfile, &st) < 0)
 				st.st_ino = -1;
+			if(stat(depfile, &st2) < 0)
+				st2.st_ino = -1;
 
-			if(f._ent->d_ino != st.st_ino) {
+			if(f._ent->d_ino != st.st_ino &&
+			   f._ent->d_ino != st2.st_ino) {
 				DEBUGP("Removing obsolete link %.*s -> %.*s\n",
 				       8, n->tupid, 8, f.filename);
 				unlinkat(f.dirfd, f.filename, 0);
